W tabeli promocja: poprawiony int na float.

+++ Supermarket2016/src/baza_supermarket_MySQL.sql
@@ -28,8 +28,8 @@ CREATE TABLE promocja
      odKiedy DATETIME , 
      doKiedy DATETIME , 
 	 idProduktu INTEGER ,
-	 staraCena INTEGER ,
-	 nowaCena INTEGER ,
+	 staraCena FLOAT ,
+	 nowaCena FLOAT ,
 	 cenaPoprzednia FLOAT ,
 	 cenaPromocyjna FLOAT
 );
